CVM: Move logic around memory faults from TDX to be shared with SNP
